Ajout de la barre de status + chgt vitesse dfl -> unsigned int

+++ V5_MVP_lecteurDiaporama/Fichiers de code/main.cpp
@@ -30,18 +30,14 @@ int main(int argc, char *argv[])
     presentation->setModele(modele);
     modele->setEtat(ModeleLecteur::Initial);
     presentation->setVue(&vueLecteur);
-
     modele->setLecteur(lecteur);
-
-
-
 
     // Connexion des signaux et des slots
 
     // Présentation -> Modele
     QObject::connect(presentation, SIGNAL(faireAvancer()), modele, SLOT(demandeAvancement()));
     QObject::connect(presentation, SIGNAL(faireReculer()), modele, SLOT(demandeReculement()));
-    QObject::connect(presentation, SIGNAL(faireChangerVitesse(float)), modele, SLOT(receptionDemandeChangementVitesse(float)));
+    QObject::connect(presentation, SIGNAL(faireChangerVitesse(unsigned int)), modele, SLOT(receptionDemandeChangementVitesse(unsigned int)));
     QObject::connect(presentation, SIGNAL(faireChangerDiapo(InfosDiaporama)), modele, SLOT(receptionDemandeChangementDiaporama(InfosDiaporama)));
 
     // Présentation -> vue
