+++ V5_MVP_lecteurDiaporama/Fichiers de code/modelelecteur.cpp
@@ -147,7 +147,7 @@ void ModeleLecteur::receptionDemandeChangementDiaporama(InfosDiaporama d)
 
 }
 
-void ModeleLecteur::receptionDemandeChangementVitesse(float pVitesse)
+void ModeleLecteur::receptionDemandeChangementVitesse(unsigned int pVitesse)
 {
     if(getEtat() != Initial)
     {
